requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
convertapi==1.8.0
distro==1.9.0
Django==4.2.14
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
django-rest-knox==5.0.1
djangorestframework==3.15.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
Markdown==3.6
openai==1.37.1
pydantic==2.8.2
pydantic_core==2.20.1
PyPDF2==2.12.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
zipp==3.19.2
